initial implementation of abbreviation replacer script. find/replace works for now, as well as enable/disable. I had to do some work to get the table to trigger an event to update the bullet optimization.
IN the process, I also found that I needed to make sure that I did
a find/replace when running updateDict() or else the keys won't be correct.

I am still not sure if I should save the keys based on their original values or the
find/replace values.

Also, there is some loop redundancy in updateDict vs updateProcessedBullets, in the
sense that both have a for loop of the exact same region... however, bulletDict is key-based,
whereas updateProcessedBullets is ordering-based. I could add a data point in bulletDict to indicate
ordering, but that will have to be done carefully, especially when new elements are added into the dictionary.

MAYBE this is where a linked list could shine... I will have to think about that.