added space optimization, bug fixes
there are a lot of thigns in this commit. First off, added a checkbox to enable or disable space optimization. That was the
main purpose of this commit/branch. In the progress of doing this, I refactored the Bullet class, primarily by separating
sentence width evaluation into its own static function. I also refactored how optimization is reported and passed on into
the window and other functions.

Bullet.EvaluateSentence takes in a sentence and width and returns an object containing two statuses: direction (whether
 to add or remove spaces) and optimization (how adequate the current sentence width is).

Bullet.optimizeSpacings now uses EvaluateSentence in two ways: initially to determine whether it needs to add or
remove spaces, and iteratively in order to check if the current iteration is adequate.

Bullet.post also uses EvaluateSentence when determining the color scheme of the displayed bullets, in specific
 cases where optimizeSentence is not evaluated For example, when space optimization is disabled. In fact, noticing
 the incorrect color scheme during implementation of space optimization disabling is what prompted me to
 refactor the Bullet class to this extent.

Some static variables were added to make status codes more understandable.

While verifying the Bullet refactoring, I noticed another minor bug-- word/lines that end with a forward slash
 (/) break differently in the browser vs in adobe PDFs. I implemented Bullet.Tweak() and Bullet.Untweak() to be
able to modify and unmodify the displayed content slightly in order to fix display compatibility issues.
In the case of the forward slash, Tweak adds a zero-width space after the slash to cause it to break properly.
Similarly, it converts dashes to non-breaking dashes (\u2011) to make it mirror what the PDF does.

I also refactored how window.BulletDict is structured. Now, within each sentence key, there is a dictionary indexed
 by widths in mm ("NNN.NNmm"). This seemed like a cleaner way to organize and evaluate optimizations.