changed where bullet optimization happens
I made the optimization routines happen on a per-bullet basis. to implement that, I made a new class called OptimizedBullet that takes care of figuring out
when to optimize and what to do once an optimization has been determined. However, the actual optimization routine is owned by the BulletComparator class.

One thing that I found through googling is the concept of debouncing as it pertains to callbacks. if I debounce my key inputs (set a callback that is interrupted if the function is called again), it
 should eliminate the slowness problem I had been experiencing earlier.