implemented abbreviations list
was able to figure out eventually how to get handsontable to work with react/babel. I think i got most of the bugs out. right now, the abbreviations table spits
out a dictionary whenever it senses a change in the abbreviations. This is sent to the BulletComparator, which takes care of actually generating the sentence replacing function callback.
I made the callback a state object so that React can tell when it changes, and update the OptimizedBullet sub-objects accordingly.