added key cacher to optimized bullet output render, added width to the key as well
to optimize performance, React uses keys to identify if a component has changed/updated. However, React complains if there are duplicate keys in a render. In
my case, I didn't care if I had duplicate keys (duplicate key == I want the exact same output anyways), but I added line numbers to my keys to suppress the complaint from React.
However, this caused performance issues with lots of bullets when the user creates a new line, since all subsequent bullets after the newline are re-rendered (keys used to be a concatenation of bullet text + line number (+ enableOptim)).

to compromise, I made a dictionary of keys, and on each individual bullet render, I check to see if the pseudo-unique key is already in the dictionary, and increment its value if so.
This prevents key collisions while also allowing for performance enhancements, since React now does not have to re-render bullets if ordering of bullets changes.