fixed issue with bullets not updating after abbreviation import
There were several things going on with this problem. Firstly, when the abbreviations table was updated, it triggered an update to the main abbrDict, but does not fundamentally change the abbrReplacer function. Since the abbrReplacer function is not updating, the bullets do not update correctly to new incoming abbreviations. I fixed this problem by explicitly abbreviating everything that comes into optimized bullet, and when those props update, optimizedBullet will update accordingly.

second, there was an issue with the heights not updating correctly for the output (to make the lines match up). this was caused in part by the new Bulma formatting-- when the abbreviations page is up, the bullets page is hidden. Bullet.evaluate() calculates things based on their appearance on the page, and since the bullets tab is hidden, all its calculations are based on zero width. To combat this, I made the entire bulletComparator conditionally appear based on what tab was selected last, instead of setting its style to 'is-hidden'. Note that I was not able to do this to the abbreviations tab due to handsontable being finicky.

There are still a few bugs that I might be seeing related to the matched line alignment.