added thesaurus replacement buttons
thesaurus replacement buttons have been added to the thesaurus list, next to each word. on the mousedown event on the plus sign icon, if the current selection is the input text box, the selected text will be replaced with whatever word was picked. the logic should automatically adjust for spaces in the selection. after replacing, the replaced word is re-selected