fixed thesaurus sporadic selection failure
issue happened when the new selection had the exact same number of characters as the old word. to force a text selection update, I set BulletEditors componentDidUpdate to look for a trigger, which is a random number generated from handleSelReplace in main.js. This should drastically reduce the likelihood of the event triggering at the wrong time.