deprecated text/setText(), fixed handleSave(), adjusted bullet output behavior
got rid of text/setText in main.js, as it was redundant with editorState. Replaced all instances of text/setText with equivalent editorState modifications.
The trickiest one was the normalize spacing one, as I had to iterate through the content blocks, find spots with /s+/g and assign corresponding selection states, and then build a new contentState based on those.

for the bullet output change, there was an issue with my shiny new bullet output renderer-- it inserted newlines in the appropriate place, but those newlines aren't actually 'real'... they are just where the break
points would normally be on the PDF form. I don't want those newlines appearing in copy/pastes, for example. To fix this, I encapuslated each line in 100% width, inline-block span elements. this caused the 'breaks'  to appear in the correct spots while not actually adding new lines to the text