switched over to non kerning and tweaked until it worked
It seems to be very clear that the Times New Roman font in the pdf forms are not kerned. Verified through the xfa-forms spec and checking the xfa data in the pdfs, as well as visually (easy to tell with combinations like 'A V')

However, when I switched over to no kerning using the CSS property font-kerning, my renderBulletText() output was not consistent with the pdf form anymore! I suspect the issue is that a PDF's input boxes can
handle slightly more width compared to an HTML input box. I ended up just adding a fudge factor of 0.55 pixels in my renderBulletText to compensate.